PlakshaChatbot/requirements.txt
google-ai-generativelanguage==0.6.15
google-generativeai==0.8.4
langchain-google-genai
protobuf<4.0.0
tenacity<9
python-dotenv
langchain-community
langchain
sentence-transformers
faiss-cpu
pypdf
